Northwind Database Connected

+++ UI/Models/ViewModel/ProductAddViewModel.cs
@@ -9,10 +9,8 @@ namespace UI.Models.ViewModel
     public class ProductAddViewModel
     {        
         public int CategoryId { get; set; }
-        //public int OrderId { get; set; }
-        //public DateTime OrderDate { get; set; }
         public string ProName { get; set; }
-        public int ProStock { get; set; }
+        public short ProStock { get; set; }
         public decimal ProPrice { get; set; }
     }
 }
